backend/requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.20.0
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.0
python-multipart==0.0.20

# AI / Langchain / Processing
langchain==0.3.23
langchain-google-genai==2.1.3
docling==2.30.0
langchain-docling==0.2.0

# Database
sqlalchemy==2.0.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.15.2
pgvector==0.4.0
tenacity==9.1.2

# Testing Dependencies
pytest==8.2.0
httpx==0.24.0
pytest-asyncio==0.26.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.0.0
bcrypt==4.0.1
